requirements.txt
Django==1.6.1
South==0.8.4
django-autocomplete-light==1.4.13
django-debug-toolbar==1.0.1
six==1.5.2
sqlparse==0.1.10
wsgiref==0.1.2
psycopg2==2.5.2
